refactor: move golang/cosmos/app/ante to golang/cosmos/ante
Matches file layout of gaia 7.
